Add files via upload
explored the queries and selected few columns like title and comments from the sqlite file.Also checked the way to extract more than 1000 posts at a time but the subreddit has only 906 posts so results did not vary but we got some recent posts. Further collected the comments for each post in comment_df and did some EDA on commensts data. Observation is 3xil3d_vinyl is the top commenter and can also see the score as 60 for few commensts so looks like some controversial discussion is in progress. Plotted a bar plot with author and score and a histogram for score frequency.Didn't try scatter plot as we had only one numeric data in comments dataframe. Also explored the n grams by installing NLTK package. Splitted the titles and removed all the whitespaces to get the most used words. Tried to clean the data by removing the stop words. first downloaded the default stop words and removed those and then added few common words as per data and some punctuations as well. Again created a plot using these most common words but now more technical terms are added to the graph.

Sentiment Analysis :-

First we downloaded the data saved in last week's assignment for data science subreddit. Then calculated the sentiment score based on keyword analysis for each title. Plotted a histogram for sentiment score and frequency.Further calculated the mean, median and standard deviation of the sentiments and the mean or average score is 0.026. Sorted the sentiments column to find out the top and bottom posts with keyword sentiment method. finally listed out the titles with top score for sentiments and 10 titles which are at the bottom of the sentiment score.

Also explored other sentiment analysis methods like textblob and VADER sentiment analysis. First installed the packages and did the necessary imports. Calculated the sentiment for each title using textBlob class polarity that gives values from -1 to +1. Plotted the histogram for tb sentiment score and frequency. The mean for this is 0.05.

Further explored the VADER sentiment method which gives us breakdown like negative, positive and neutral. ALso calculated the VADER sentiment score "compound" for each title.
The VADER sentiment analysis seems to be the bext model for this data as it gives the average of 0.078. Then used the data to get the post and comment dataframe for analysis. Used textblob method to get the sentiment score for text columns in post_df.Also applied textblob method to get the sentiment analysis for comments df. Worked on getting the sentiment score for each post using text blob method. Took the mean of sentiment score for all the comments in each post to get the sentiment score for the posts. The result is saved in post_sentiment_df data frame. Also listed out the top and bottom posts based on sentiment score. Further plotted the histogram for sentiment average and frequency. Worked on subjectivity and polarity of sentiments using text blob method. Used the saved data frame df_copy for this analysis. Got the subjectivity of each title using the function with textBlob class. Also calculated the polarity for each title. Further checked on the group polarity. The function is defined to get the values as Negative (for score <0), Neutral (for score =0) and Positive (for score >0) and saved the result in “Insight” column. Visualised the Insight plot and observation is that maximum titles have neutral score followed by positive in terms of sentiments.

Topic Modelling - Wordcloud and LDA model :- Explored the Topic modelling on this data. First checked the word cloud for all the titles to get the most used words that make sense (after removing the stopwords). Most words are fine and technical terms as per the topic. Then explored the Genesis package for LDA modelling. Took a sample of around 100 records to work on. Dropped all the extra columns and have taken the text columns for modelling. Cleaning the text column of data by removing all the stop words, punctuations and converting the text to small case. Then converted the same text to dictionary and corpus. Trained the model further for 10 topics to get the most related words for the topic and also how these are related with each other.
Installed pyLDAvis package to visualise the result of this LDA model. Got the graph that gives most related words for each topic and the Intertopic distance map that gives the understanding how much each topic is related to another. ALso attached the html file generated using pyLDAvis package.